requirements.txt
streamlit==1.32.0
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
Pillow==10.2.0
google-generativeai==0.3.2
python-multipart==0.0.9
requests==2.31.0 